pyproject.toml
[project]
name = "ai-transcribe"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrejus",email = "anba8005@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch (>=2.6.0,<3.0.0)",
    "torchaudio (>=2.6.0,<3.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "pyannote-audio (>=3.3.2,<4.0.0)",
    "speechbrain (>=1.0.2,<2.0.0)",
    "faster-whisper (>=1.1.1,<2.0.0)",
    "ctranslate2 (==4.4.0)",
    "pydub (>=0.25.1,<0.26.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
run = "python src/main.py"